# add mkldnn softmax_output