--- a/src/operator/softmax_output-inl.h
+++ b/src/operator/softmax_output-inl.h
@@ -88,6 +88,7 @@ struct SoftmaxOutputParam : public dmlc::Parameter<SoftmaxOutputParam> {
               "one-hot encoding of the gold label and distributed uniformly to"
               "all other labels.");
   };
+
   bool operator==(const SoftmaxOutputParam& other) const {
     return this->grad_scale == other.grad_scale &&
     this->ignore_label == other.ignore_label &&
@@ -277,32 +278,28 @@ class SoftmaxOutputOp : public Operator {
   SoftmaxOutputParam param_;
 };  // class SoftmaxOutputOp
 
-// Decalre Factory function, used for dispatch specialization
-template<typename xpu>
-Operator* CreateOp(SoftmaxOutputParam param, int dtype);
-
 template<typename xpu>
 void SoftmaxOutputCompute(const nnvm::NodeAttrs& attrs,
-                      const OpContext& ctx, const std::vector<TBlob>& inputs,
-                      const std::vector<OpReqType>& req,
-                      const std::vector<TBlob>& outputs) {
+                          const OpContext& ctx, const std::vector<TBlob>& inputs,
+                          const std::vector<OpReqType>& req,
+                          const std::vector<TBlob>& outputs) {
   const SoftmaxOutputParam &param = nnvm::get<SoftmaxOutputParam>(attrs.parsed);
   const std::vector<TBlob> no_use_but_adapt_origin_api;
   CHECK_EQ(inputs.size(), 2U);
   std::vector<TBlob> in_data(inputs.begin(),
                              inputs.begin() + softmaxout_enum::kLabel);
   MSHADOW_REAL_TYPE_SWITCH(inputs[softmaxout_enum::kData].type_flag_, DType, {
-  SoftmaxOutputOp<xpu, DType>op(param);
-  op.Forward(ctx, inputs, req, outputs, no_use_but_adapt_origin_api);
+    SoftmaxOutputOp<xpu, DType> op(param);
+    op.Forward(ctx, inputs, req, outputs, no_use_but_adapt_origin_api);
   });
 }
 
 template<typename xpu>
 void SoftmaxOutputGradCompute(const nnvm::NodeAttrs& attrs,
-                               const OpContext& ctx,
-                               const std::vector<TBlob>& inputs,
-                               const std::vector<OpReqType>& req,
-                               const std::vector<TBlob>& outputs) {
+                              const OpContext& ctx,
+                              const std::vector<TBlob>& inputs,
+                              const std::vector<OpReqType>& req,
+                              const std::vector<TBlob>& outputs) {
   const SoftmaxOutputParam& param = nnvm::get<SoftmaxOutputParam>(attrs.parsed);
   const std::vector<TBlob> no_use_but_adapt_origin_api;
   CHECK_EQ(inputs.size(), 2U);
@@ -314,8 +311,8 @@ void SoftmaxOutputGradCompute(const nnvm::NodeAttrs& attrs,
   const std::vector<TBlob> &in_grad = outputs;
 
   MSHADOW_REAL_TYPE_SWITCH(dtype, DType, {
-  SoftmaxOutputOp<xpu, DType>op(param);
-  op.Backward(ctx, out_grad, in_data, out_data, req, in_grad, no_use_but_adapt_origin_api);
+    SoftmaxOutputOp<xpu, DType> op(param);
+    op.Backward(ctx, out_grad, in_data, out_data, req, in_grad, no_use_but_adapt_origin_api);
   });
 }