… src/operator/softmax_output.cu