@pengzhao-intel 2018-12-22
@rongzha1 please help fix the lint issue 

> Warning, treated as error:
> 
> /work/mxnet/python/mxnet/ndarray/__init__.py:docstring of mxnet.ndarray.SoftmaxOutput:30: WARNING: Bullet list ends without a blank line; unexpected unindent.