@rongzha1 2018-12-26
> @rongzha1 could you post some performance changes by this PR?

for 1024*256 input, performance speedup 2.75

env: skx-8180, 1socket 28 core, 
model : wide_and_deep model


old (ms) | opt(ms) | speedup
-- | -- | --
0.1890553 | 0.0687512 | 2.7498492