@pengzhao-intel 2019-01-02
@rongzha1 please rebase the code and make the CI pass :)