@rongzha1 2019-01-03
please help to review and merge to the master branch. @eric-haibin-lin @zheng-da @azai91