@pengzhao-intel 2019-02-02
@szha could you help take a look for the API change?